requirements.txt
azure-functions==1.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
idna==3.3
imageio==2.19.2
numpy==1.22.3
Pillow==9.1.1
pyspng==0.1.0
opencv-python==4.5.5.64
requests==2.27.1
scipy==1.8.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
typing_extensions==4.2.0
urllib3==1.26.9
